requirements.txt
# Core dependencies for DialoGPT RAG system (lightweight)
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pandas>=1.3.0
numpy>=1.21.0
pickle-mixin>=1.0.2

# Optional dependencies for time series forecasting
prophet>=1.1.0

# For better performance (optional)
accelerate>=0.20.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
huggingface-hub>=0.36.0
